+++ src/main/java/org/jsoup/select/QueryParser.java
@@ -145,21 +145,21 @@ public class QueryParser {
 
     private String consumeSubQuery() {
         StringBuilder sq = StringUtil.borrowBuilder();
-        boolean seenNonCombinator = false; // eat until we hit a combinator after eating something else
+        boolean seenClause = false; // eat until we hit a combinator after eating something else
         while (!tq.isEmpty()) {
+            if (tq.matchesAny(Combinators)) {
+                if (seenClause)
+                    break;
+                sq.append(tq.consume());
+                continue;
+            }
+            seenClause = true;
             if (tq.matches("("))
                 sq.append("(").append(tq.chompBalanced('(', ')')).append(")");
             else if (tq.matches("["))
                 sq.append("[").append(tq.chompBalanced('[', ']')).append("]");
-            else if (tq.matchesAny(Combinators))
-                if (seenNonCombinator)
-                    break;
-                else
-                    sq.append(tq.consume());
-            else {
-                seenNonCombinator = true;
+            else
                 sq.append(tq.consume());
-            }
         }
         return StringUtil.releaseBuilder(sq);
     }